#171 [fix] 카테고리별 데이터셋 개수 조회 응답에서 개수를 int에서 Long으로 변환한다.

+++ src/main/java/com/dataracy/modules/dataset/adapter/web/response/CountDataGroupWebResponse.java
@@ -3,5 +3,5 @@ package com.dataracy.modules.dataset.adapter.web.response;
 public record CountDataGroupWebResponse(
         Long topicId,
         String topicLabel,
-        int count
+        Long count
 ) {}
