+++ src/main/java/com/dataracy/modules/dataset/application/dto/response/CountDataGroupResponse.java
@@ -3,5 +3,5 @@ package com.dataracy.modules.dataset.application.dto.response;
 public record CountDataGroupResponse(
         Long topicId,
         String topicLabel,
-        int count
+        Long count
 ) {}
